refactor(api): Phase 3b — field privacy / non_exhaustive / sealed-trait decisions (closes #190, #222)
Phase 3b of the API hygiene playbook: systematically audit and document
decisions for §3.4 (public struct field discipline), §3.5 (#[must_use]
opportunities), §3.6 (#[non_exhaustive] on enums/structs), and §3.7
(sealed-trait pattern) across the five audit-target crates.

Strategy: per-scope decision records, not per-item.  Quantifying the
audit load (84 + 69 = 153 structs; 10 + 35 = 45 enums in uffs-mft +
uffs-core alone) made it clear that per-item records would produce ~200
nearly-identical doc-comment blocks.  The execution therefore landed:

  * crate-level policy in uffs-mft::lib.rs, uffs-core::lib.rs, and
    uffs-client::lib.rs;
  * type-level records for genuine exceptions (UffsClient /
    UffsClientSync / UffsMcpServer private-field success stories;
    FormatRow + FileReader sealed-trait kept-open decisions).

§3.4 outcome (per-crate):
  - uffs-format: 4 type records (OutputConfig, DriveFooterContext,
    OutputColumn, FormatRow).
  - uffs-mcp: 3 type records (McpConfig + HttpGatewayConfig pub fields;
    UffsMcpServer private fields with smart constructors).
  - uffs-client: crate-level policy covering protocol DTOs, schema
    metadata, UffsClient / UffsClientSync, and ClientError (single
    consolidated record at lib.rs rather than per-module).
  - uffs-mft: crate-level policy covering 3 categories — NTFS zerocopy
    types, index/record DTOs, config option structs.
  - uffs-core: crate-level policy covering 4 categories — aggregate
    specs/results, search backend value types, compact-cache headers,
    pattern/trigram/bloom data structures.

§3.5 outcome: 779 #[must_use] attrs already present across the 5 audit
crates; sub-phase audits found no missing candidates.  Recorded "no
fresh candidates" outcome.

§3.6 outcome: ClientError (uffs-client::error) is the single new
#[non_exhaustive] annotation.  Verified safe — both external consumers
in uffs-cli use `if let Some(ClientError::X { … }) = …` or variant
construction `ClientError::DaemonNeedsElevation { … }` (non-exhaustive
on enum, not on variants, permits this).  All other enums kept
exhaustive: closed type-code enums by NTFS spec (uffs-mft), state-
machine / dispatch enums consumed by 100s of exhaustive match arms
(uffs-core, uffs-client), and config DTOs (uffs-format, uffs-mcp).

§3.7 outcome: 2 pub traits documented as kept-open with explicit
rationale: FormatRow (uffs-format) — cross-crate impls in
uffs-client::SearchRow + uffs-core::DisplayRow; FileReader
(uffs-core::aggregate::verify) — deliberate DI seam between daemon
production reader and test mock reader.

Two file splits to honour the 800-LOC policy after doc additions:
  - connect_sync.rs (800 → 696 LOC): auto_start_daemon /
    is_process_alive / is_daemon_process moved to new sibling
    `connect_sync_autostart.rs` (PID-file freshness → liveness check →
    daemon-identity verification — a cohesive 3-fn cluster only used
    by the sync client's reconnect loop).
  - protocol/mod.rs (793 → 603 LOC): aggregate wire DTOs
    (AggregateSpecWire / AggregateResultWire / StatsWire / BucketWire /
    SampleRowWire / DrilldownWire) moved to new sibling
    `protocol/aggregate_wire.rs` and re-exported so existing
    `uffs_client::protocol::AggregateSpecWire` import paths keep working.

Strict-lint posture honoured throughout: zero new #[allow(...)]
directives, no lint-disable flags, no #[cfg(...)] dead-code workarounds;
2 intra-doc-link adjustments to replace private-item links with plain
code spans (surgical, in-place).

Verification:
  - cargo check --workspace --all-targets: 0 warnings
  - RUSTDOCFLAGS=-Dwarnings cargo doc --workspace --no-deps: 0 errors
  - cargo nextest run --workspace: 1,710 / 1,710 pass, 14 skipped
  - file-size policy gate: green (no new MISSING_EXCEPTION)
  - just lint-prod: green

20 files changed, +668 / -308 net.

Closes #190 (Phase 3 umbrella).
Closes #222 (Phase 3b follow-up).
