+++ crates/uffs-format/src/footer.rs
@@ -35,6 +35,22 @@ use std::io::{self, Write};
 /// `uffs-daemon::handler` (fast path) can share the exact same struct.
 /// Borrowed slices / strs keep the type free of allocations — the
 /// footer writer only reads, never takes ownership.
+///
+/// # Field discipline (Phase 3b §3.4)
+///
+/// All three fields are `pub` because they are **required positional
+/// inputs** to [`write_legacy_drive_footer`].  A builder pattern would
+/// add lifetime-parameter friction (`DriveFooterContextBuilder<'a>`)
+/// without changing the required-vs-optional nature of any field.
+///
+/// # `#[non_exhaustive]` decision (Phase 3b §3.6)
+///
+/// **Kept exhaustive.**  This is a borrowed-data DTO whose three
+/// fields are all required arguments to the writer; future growth
+/// would mean a new required argument (which is a breaking change
+/// regardless of `#[non_exhaustive]`).  Both call sites
+/// (`uffs_cli::commands::output::parity::write_legacy_drive_footer`
+/// and `uffs_daemon::handler_blob`) live in the same workspace.
 #[derive(Debug, Clone, Copy)]
 pub struct DriveFooterContext<'a> {
     /// Drive letters the search targeted (e.g. `['C', 'D']`).  When
